Extracted job_status enum in separate file.
