+++ python/tests/res/job_queue/test_job_queue.py
@@ -5,7 +5,7 @@ from tests import ResTest
 class JobQueueTest(ResTest):
 
     def testStatusEnum(self):
-        source_path = "libjob_queue/include/ert/job_queue/queue_driver.h"
+        source_path = "libjob_queue/include/ert/job_queue/job_status.h"
         self.assertEnumIsFullyDefined(JobStatusType, "job_status_type", source_path)
